implement support for WCMP2
