+++ README.md
@@ -9,6 +9,7 @@ WIS](https://community.wmo.int/activity-areas/wis/wis-overview) Core Metadata
 Profile (WCMP).
 
 - validation against [WCMP 1.3](http://wis.wmo.int/2013/metadata/version_1-3-0/WMO_Core_Metadata_Profile_v1.3_Part_1.pdf), specifically [Part 2](http://wis.wmo.int/2013/metadata/version_1-3-0/WMO_Core_Metadata_Profile_v1.3_Part_2.pdf), Section 2, implementing an executable test suite against the ATS
+- validation against [WCMP 2 (draft)](https://github.com/wmo-im/wcmp2), specifically [Annex A: Conformance Class Abstract Test Suite](https://wmo-im.github.io/wcmp2/standard/wcmp2-DRAFT.html#_conformance_class_abstract_test_suite_normative), implementing an executable test suite against the ATS
 - quality assessement against the [WCMP Key Performance Indicators](https://community.wmo.int/activity-areas/wis/wis-metadata-kpis)
 
 ## Installation
@@ -47,41 +48,44 @@ pywcmp bundle sync
 
 # abstract test suite
 
-# validate metadata against abstract test suite (file on disk)
-pywcmp ats validate --file /path/to/file.xml
+# validate WCMP 1.3 metadata against abstract test suite (file on disk)
+pywcmp ats validate /path/to/file.xml
 
-# validate metadata against abstract test suite (URL)
-pywcmp ats validate --url http://example.org/path/to/file.xml
+# validate WCMP 1.3 metadata against abstract test suite (URL)
+pywcmp ats validate https://example.org/path/to/file.xml
+
+# validate WCMP 2 metadata against abstract test suite (URL)
+pywcmp ats validate https://example.org/path/to/file.json
 
 # adjust debugging messages (CRITICAL, ERROR, WARNING, INFO, DEBUG) to stdout
-pywcmp ats validate --url http://example.org/path/to/file.xml --verbosity DEBUG
+pywcmp ats validate https://example.org/path/to/file.xml --verbosity DEBUG
 
 # write results to logfile
-pywcmp ats validate --url http://example.org/path/to/file.xml --verbosity DEBUG --logfile /tmp/foo.txt
+pywcmp ats validate https://example.org/path/to/file.json --verbosity DEBUG --logfile /tmp/foo.txt
 
 # key performance indicators
 
 # all key performance indicators at once # note: running KPIs automatically runs the ats
-pywcmp kpi validate --url http://example.org/path/to/file.xml --verbosity DEBUG
+pywcmp kpi validate https://example.org/path/to/file.xml --verbosity DEBUG
 
 # all key performance indicators at once, in summary
-pywcmp kpi validate --url http://example.org/path/to/file.xml --verbosity DEBUG --summary
+pywcmp kpi validate https://example.org/path/to/file.xml --verbosity DEBUG --summary
 
 # all key performance indicators at once, with scoring rubric grouping
-pywcmp kpi validate --url http://example.org/path/to/file.xml --verbosity DEBUG --group
+pywcmp kpi validate https://example.org/path/to/file.xml --verbosity DEBUG --group
 
 # selected key performance indicator
-pywcmp kpi validate --kpi 4 -f /path/to/file.xml -v INFO
+pywcmp kpi validate --kpi 4 /path/to/file.xml -v INFO
 
-# WIS topic hierarchies
+# WIS2 topic hierarchies
 
-# validate a WIS 2.0 topic hierarchy
+# validate a WIS2 topic hierarchy
 pywcmp topics validate origin.a.wis2.can
 
-# validate a partial WIS 2.0 topic hierarchy
+# validate a partial WIS2 topic hierarchy
 pywcmp topics validate wis2.can --fuzzy
 
-# list children of a given WIS 2.0 topic hierarchy level
+# list children of a given WIS2 topic hierarchy level
 pywcmp topics list wis2.a
 
 ```
@@ -98,15 +102,15 @@ pywcmp topics list wis2.a
 >>> # test a URL
 >>> from urllib2 import urlopen
 >>> from StringIO import StringIO
->>> content = StringIO(urlopen('http://....').read())
+>>> content = StringIO(urlopen('https://....').read())
 >>> exml = etree.parse(content)
 >>> ts = ats.WMOCoreMetadataProfileTestSuite13(exml)
 >>> ts.run_tests()  # raises ValueError error stack on exception
->>> # handle ats.TestSuiteError
->>> # ats.TestSuiteError.errors is a list of errors
+>>> # handle pywcmp.errors.TestSuiteError
+>>> # pywcmp.errors.TestSuiteError.errors is a list of errors
 >>> try:
 ...    ts.run_tests()
-... except ats.TestSuiteError as err:
+... except pywcmp.errors.TestSuiteError as err:
 ...    print('\n'.join(err.errors))
 >>> ...
 >>> # test KPI
