+++ pywcmp/util.py
@@ -26,28 +26,37 @@
 #
 ###############################################################################
 
+import json
 import logging
-import os
 from pathlib import Path
 import ssl
 import sys
-from datetime import datetime, timezone, timedelta
-from dateutil.parser import parse
 from urllib.error import URLError
 from urllib.request import urlopen
 from urllib.parse import urlparse
 
+from spellchecker import SpellChecker
 from lxml import etree
 
 LOGGER = logging.getLogger(__name__)
+THISDIR = Path(__file__).parent.resolve()
 
-NAMESPACES = {
-    'gco': 'http://www.isotc211.org/2005/gco',
-    'gmd': 'http://www.isotc211.org/2005/gmd',
-    'gml': 'http://www.opengis.net/gml/3.2',
-    'gmx': 'http://www.isotc211.org/2005/gmx',
-    'xlink': 'http://www.w3.org/1999/xlink'
-}
+
+def check_spelling(text: str) -> list:
+    """
+    Helper function to spell check a string
+
+    :returns: `list` of unknown / misspelled words
+    """
+
+    LOGGER.debug(f'Spellchecking {text}')
+    spell = SpellChecker()
+
+    dictionary = THISDIR / 'dictionary.txt'
+    LOGGER.debug(f'Loading custom dictionary {dictionary}')
+    spell.word_frequency.load_text_file(f'{dictionary}')
+
+    return list(spell.unknown(spell.split_words(text)))
 
 
 def get_cli_common_options(function):
@@ -66,124 +75,6 @@ def get_cli_common_options(function):
     return function
 
 
-def get_codelists():
-    """
-    Helper function to assemble dict of ISO and WMO codelists
-
-    :param authority: code list authority (iso or wmo)
-
-    :returns: `dict` of ISO and WMO codelists
-
-    """
-
-    codelists = {}
-    userdir = get_userdir()
-
-    codelist_files = {
-        'iso': f'{userdir}/wcmp-1.3/schema/resources/Codelist/gmxCodelists.xml',
-        'wmo': f'{userdir}{os.sep}wcmp-1.3{os.sep}WMOCodeLists.xml'
-    }
-
-    for key, value in codelist_files.items():
-        codelists[key] = {}
-        xml = etree.parse(value)
-        for cld in xml.xpath('gmx:codelistItem/gmx:CodeListDictionary', namespaces=NAMESPACES):
-            identifier = cld.get(nspath_eval('gml:id'))
-            codelists[key][identifier] = []
-            for centry in cld.findall(nspath_eval('gmx:codeEntry/gmx:CodeDefinition/gml:identifier')):
-                codelists[key][identifier].append(centry.text)
-
-    return codelists
-
-
-def get_string_or_anchor_value(parent) -> list:
-    """
-    Returns list of strings (texts) from CharacterString or Anchor child elements of the given element
-
-    :param parent : The element to check
-    """
-    values = []
-    value_elements = parent.findall(nspath_eval('gco:CharacterString')) + parent.findall(nspath_eval('gmx:Anchor'))
-    for element in value_elements:
-        values.append(element.text)
-    return values
-
-
-def get_string_or_anchor_values(parent_elements: list) -> list:
-    """
-    Returns list of strings (texts) from CharacterString or Anchor child elements of given parent_elements
-
-    :param parent_elements : List of parent elements of the CharacterString or Anchor to read.
-    """
-    values = []
-    for parent in parent_elements:
-        values += get_string_or_anchor_value(parent)
-    return values
-
-
-def get_keyword_info(main_keyword_element) -> tuple:
-    """
-    Returns tuple with keywords, type value(s) and thesaurus(es) for given "MD_Keywords" element
-
-    :param main_keyword_element : The element to check
-    """
-
-    keywords = main_keyword_element.findall(nspath_eval('gmd:keyword'))
-    type_element = get_codelist_values(main_keyword_element.findall(nspath_eval('gmd:type/gmd:MD_KeywordTypeCode')))
-    thesauruses = main_keyword_element.findall(nspath_eval('gmd:thesaurusName/gmd:CI_Citation/gmd:title'))
-    return keywords, type_element, thesauruses
-
-
-def get_codelist_values(elements: list) -> list:
-    """
-    Returns list of code list values as strings for all elements (except the ones with no value)
-    The value can be in the element attribute or text node.
-
-    :param elements : The elements to check
-    """
-    values = []
-    for element in elements:
-        value = element.get('codeListValue')
-        if value is None:
-            value = element.text
-        if value is not None:
-            values.append(value)
-    return values
-
-
-def parse_time_position(element) -> datetime:
-    """
-    Returns datetime extracted from the given GML element or None if parsing failed.
-    The parsing is rather benevolent here to allow mixing of "Zulu" and "naive" time strings (and other oddities),
-    in the hope that all meteorological data refer to UTC.
-
-    :param element : XML / GML element (e.g. gml:beginPosition)
-    """
-    indeterminate_pos = element.get('indeterminatePosition')
-    if indeterminate_pos is not None:
-        if indeterminate_pos in ["now", "unknown"]:
-            return datetime.now(timezone.utc)
-        elif indeterminate_pos == "before":
-            return datetime.now(timezone.utc) - timedelta(hours=24)
-        elif indeterminate_pos == "after":
-            return datetime.now(timezone.utc) + timedelta(hours=24)
-        else:
-            LOGGER.debug(f'Time point has unexpected value of indeterminatePosition: {indeterminate_pos}')
-    elif element.text is not None:
-        text_to_parse = element.text
-        if text_to_parse.endswith('Z'):
-            text_to_parse = text_to_parse[0:-1]
-
-        try:
-            dtg = parse(text_to_parse, fuzzy=True, ignoretz=True).replace(tzinfo=timezone.utc)
-            return dtg
-        except Exception as err:
-            msg = f'Invalid time string: {err}'
-            LOGGER.debug(msg)
-
-    return None
-
-
 def get_userdir() -> str:
     """
     Helper function to get userdir
@@ -194,27 +85,7 @@ def get_userdir() -> str:
     return Path.home() / '.pywcmp'
 
 
-def nspath_eval(xpath: str) -> str:
-    """
-    Return an etree friendly xpath based expanding namespace
-    into namespace URIs
-
-    :param xpath: xpath string with namespace prefixes
-
-    :returns: etree friendly xpath
-    """
-
-    out = []
-    for chunk in xpath.split('/'):
-        if ':' in chunk:
-            namespace, element = chunk.split(':')
-            out.append(f'{{{NAMESPACES[namespace]}}}{element}')
-        else:
-            out.append(chunk)
-    return '/'.join(out)
-
-
-def setup_logger(loglevel: str = None, logfile: str = None):
+def setup_logger(loglevel: str = None, logfile: str = None) -> None:
     """
     Setup logging
 
@@ -293,7 +164,7 @@ def check_url(url: str, check_ssl: bool, timeout: int = 30) -> dict:
     }
 
     try:
-        if check_ssl is False:
+        if not check_ssl:
             LOGGER.debug('Creating unverified context')
             result['ssl'] = False
             context = ssl._create_unverified_context()
@@ -309,7 +180,7 @@ def check_url(url: str, check_ssl: bool, timeout: int = 30) -> dict:
         LOGGER.debug(f'Other error: {err}')
         LOGGER.debug(err)
 
-    if response is None and check_ssl is True:
+    if response is None and check_ssl:
         return check_url(url, False)
 
     if response is not None:
@@ -322,51 +193,46 @@ def check_url(url: str, check_ssl: bool, timeout: int = 30) -> dict:
             result['mime-type'] = response.headers.get_content_type()
         else:
             result['accessible'] = True
-        if parsed_uri.scheme in ('https') and check_ssl is True:
+        if parsed_uri.scheme in ('https') and check_ssl:
             result['ssl'] = True
     else:
         result['accessible'] = False
     return result
 
 
-def validate_iso_xml(xml):
+def parse_wcmp(content: str) -> list:
     """
-    Perform XML Schema validation of ISO XML Metadata
+    Parse a buffer into an etree ElementTree (WCMP1) or JSON dict (WCMP2)
 
-    :param xml: file or string of XML
+    :param content: str of JSON or XML
 
-    :returns: `bool` of whether XML validates ISO schema
-    """
-
-    userdir = get_userdir()
-    if not os.path.exists(userdir):
-        raise IOError(f'{userdir} does not exist')
-    if isinstance(xml, str):
-        xml = etree.fromstring(xml)
-    xsd = os.path.join(userdir, 'wcmp-1.3', 'iso-all.xsd')
-    LOGGER.debug(f'Validating {xml} against schema {xsd}')
-    schema = etree.XMLSchema(etree.parse(xsd))
-    schema.assertValid(xml)
-
-
-def parse_wcmp(content):
-    """
-    Parse a buffer into an etree ElementTree
-
-    :param content: str of xml content
-
-    :returns: `lxml.etree._ElementTree` object of WCMP
+    :returns: list of:
+              -  `lxml.etree._ElementTree` or `dict` object of WCMP
+              -  WCMP version
     """
 
-    try:
-        exml = etree.parse(content)
-    except etree.XMLSyntaxError as err:
-        LOGGER.error(err)
-        raise RuntimeError('Syntax error')
+    wcmp_version_guess = 1
 
-    root_tag = exml.getroot().tag
+    if content.strip().endswith('.json'):
+        LOGGER.debug('Attempting to parse as JSON')
+        try:
+            with open(content) as fh:
+                data = json.load(fh)
+            wcmp_version_guess = 2
+        except RuntimeError as err:
+            LOGGER.error(err)
+            raise RuntimeError('Syntax error')
+    else:
+        LOGGER.debug('Attempting to parse as XML')
+        try:
+            data = etree.parse(content)
+        except etree.XMLSyntaxError as err:
+            LOGGER.error(err)
+            raise RuntimeError('Syntax error')
 
-    if root_tag != '{http://www.isotc211.org/2005/gmd}MD_Metadata':
-        raise RuntimeError('Does not look like a WCMP document!')
+        root_tag = data.getroot().tag
 
-    return exml
+        if root_tag != '{http://www.isotc211.org/2005/gmd}MD_Metadata':
+            raise RuntimeError('Does not look like a WCMP document!')
+
+    return data, wcmp_version_guess
